# sync-server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fractonica API
  version: current
  summary: Server-authoritative journal, event, and synchronization API.
  description: |
    The Fractonica API is the canonical server contract.
    It is multi-user, cursor-synchronized, and intended for the first-party iOS
    client as well as user-authorized automation agents.

    Authentication uses either a short-lived JWT access token (`JwtBearer`) or
    a user-issued API key (`ApiKeyBearer`) in the `Authorization: Bearer ...`
    header. API keys are bound to one user and one logical device. Operations
    carrying `x-required-scopes` enforce those scopes for API-key callers; JWT
    sessions act with the authenticated user's full authority.

    Production deployments must expose the API only over HTTPS/TLS and reject
    plain HTTP at the edge. Passwords and bearer credentials are otherwise
    observable in transit. Plain HTTP is for loopback development only.

    Records are public by default. Public record payloads are readable and may
    be rendered from a template. Private record payloads and private
    media are encrypted by the client before upload; the server stores only an
    authenticated ciphertext envelope and intentionally cannot render, index,
    or embed their plaintext.

    All timestamps are RFC 3339 instants. Record API identifiers are globally
    unique five-character Base64URL strings; their owner-only `originId` UUIDs
...
    are UUIDs.
...

...
    immediately retrying.

...
  license:
...
servers:
...
security:
  - JwtBearer: []
  - ApiKeyBearer: []
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Authentication]
      summary: Register a user and start a session
      description: |
        Availability depends on the deployment registration mode (`open`,
        `invite`, or `closed`). Passwords are never used directly as private
        record encryption keys.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              login: aurora
              password: correct-horse-battery-staple
              displayName: Aurora Observer
              inviteCode: FRACTONICA-2026
      responses:
        '201':
          description: User and initial session created.
          headers:
            Location:
              description: URL of the new user resource.
              schema:
                type: string
                format: uri-reference
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/login:
    post:
      operationId: loginUser
      tags: [Authentication]
      summary: Exchange login and password for a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              login: aurora
              password: correct-horse-battery-staple
      responses:
        '200':
          description: Login succeeded.
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [Authentication]
      summary: Rotate a refresh token and issue a new access token
      description: |
        Refresh tokens are opaque, single-use rotation credentials. Reuse of a
        previously rotated token revokes its token family.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token rotation succeeded.
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [Authentication]
      summary: Revoke the current refresh-token family
      security:
        - JwtBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Session revoked. The operation is idempotent.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /me:
    get:
      operationId: getCurrentUser
      tags: [Authentication]
      summary: Get the authenticated user
      security:
        - JwtBearer: []
        - ApiKeyBearer: []
      responses:
        '200':
          description: Current user profile.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    patch:
      operationId: updateCurrentUser
      tags: [Authentication]
      summary: Update authenticated user preferences
      description: |
        JWT-only. Updates the Saros-series anchor used by temporal clocks and
        record presentation. Supply the strong ETag returned by `GET /me`.
      security:
        - JwtBearer: []
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User preferences updated.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived Ed25519-signed access JWT issued by login or refresh. The
        token includes `iss`, `aud`, `sub`, `exp`, and `kid` claims.
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: Fractonica API key
      description: |
        User-issued secret beginning with `exk_`. Keys are stored hashed,
        device-bound, scoped, independently expirable, and revocable. The
        plaintext secret appears only in the initial creation response; a replay
        returns the created `apiKeyId`, never the secret.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag of the revision being mutated.
      schema:
        type: string
        minLength: 3
        maxLength: 200
      example: '"record-aB_9Z-r4"'
...
  headers:
    ETag:
      description: Strong opaque validator for the returned resource revision.
      schema:
        type: string
      example: '"record-aB_9Z-r4"'
    RequestId:
      description: Server-assigned request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: The request syntax, parameter combination, or cursor is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.exeligmos.app/problems/invalid-request
            title: Invalid request
            status: 400
            detail: The occurredAfter timestamp must precede occurredBefore.
            code: invalid_request
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    Unauthorized:
      description: Authentication is missing, expired, revoked, or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="exeligmos"
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the scope or ownership required for this operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: The mutation conflicts with current state or idempotency history.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PreconditionFailed:
      description: "`If-Match` does not match the current resource revision."
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    UnprocessableContent:
      description: The request is syntactically valid but violates domain constraints.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    ServiceUnavailable:
      description: A database statement or lock exceeded its safety deadline; the operation may be retried.
      headers:
        Retry-After:
          description: Minimum number of seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
          example: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:exeligmos:problem:database-timeout
            title: Service Unavailable
            status: 503
            detail: The database could not complete the request within its safety deadline.
            code: database_timeout
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    TooManyRequests:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details with stable application extensions.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type URI.
        title:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: true
    FieldError:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: string
          description: JSON Pointer to the invalid value.
          maxLength: 500
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        message:
          type: string
          minLength: 1
          maxLength: 1000
      additionalProperties: false
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/FieldError'
...
    RegisterRequest:
      type: object
      required: [login, password]
      properties:
        login:
          $ref: '#/components/schemas/LoginName'
        password:
          $ref: '#/components/schemas/Password'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        inviteCode:
          type: string
          minLength: 1
          maxLength: 200
      additionalProperties: false
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          $ref: '#/components/schemas/LoginName'
        password:
          $ref: '#/components/schemas/Password'
      additionalProperties: false
    LoginName:
      type: string
      description: Case-insensitive account login stored in normalized form.
      minLength: 3
      maxLength: 64
      pattern: '^[A-Za-z0-9][A-Za-z0-9._-]*$'
...
    Password:
      type: string
      format: password
      minLength: 12
      maxLength: 1024
      writeOnly: true
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 2048
          writeOnly: true
      additionalProperties: false
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 2048
          writeOnly: true
      additionalProperties: false
    AuthSession:
      type: object
      required:
        - tokenType
        - accessToken
        - expiresIn
        - refreshToken
        - refreshExpiresIn
        - user
      properties:
        tokenType:
          type: string
          const: Bearer
        accessToken:
          type: string
          minLength: 32
          readOnly: true
        expiresIn:
          type: integer
          description: Access-token lifetime in seconds.
          minimum: 1
        refreshToken:
          type: string
          minLength: 32
          readOnly: true
        refreshExpiresIn:
          type: integer
          description: Refresh-token lifetime in seconds.
          minimum: 1
        user:
          $ref: '#/components/schemas/User'
      additionalProperties: false
    User:
      type: object
      required: [id, login, displayName, sarosAnchor, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        login:
          $ref: '#/components/schemas/LoginName'
        displayName:
          type: string
          minLength: 1
          maxLength: 120
        sarosAnchor:
          $ref: '#/components/schemas/SarosAnchor'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    UpdateUserRequest:
      type: object
      required: [sarosAnchor]
      properties:
        sarosAnchor:
          $ref: '#/components/schemas/SarosAnchor'
      additionalProperties: false
    SarosAnchor:
      type: integer
      minimum: 1
      maximum: 180
      default: 141
      description: Saros series used as the user's temporal pulse anchor.
...